Move SmartQuotes to sphinx.transforms.post_transforms
As prepartion for refactoring, this moves SmartQuotes to
post_transforms package (no modification).

related change: NoUri exception is moved to errors package
(to avoid circular import)
